content/base/test/test_CrossSiteXHR2.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Cross Site XMLHttpRequest with IDN</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript"
    src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>

<iframe id="inner" src="http://sub1.xn--lt-uia.example.org:8000/tests/content/base/test/file_CrossSiteXHR2_inner.html">
</iframe>

<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

var t = setInterval(doCheck, 300);
function doCheck()
{
  if (location.hash) {
    clearInterval(t);
    is(location.hash, "#done", "failed to run");
    SimpleTest.finish();
  }
}

</script>
</pre>
</body>
</html>
